fix: fixed delete user endpoint being a GET request (lol)

+++ src/app/routers/users.py
@@ -1,4 +1,4 @@
-from flask import Blueprint, redirect, render_template
+from flask import Blueprint, redirect, render_template, request
 from flask_babel import format_datetime
 from flask_login import current_user, login_required
 
@@ -29,6 +29,6 @@ def delete_user(user_id):
     Deletes the user. Does nothing on attempt to delete self. Requires login
     """
 
-    if current_user.id != user_id:
+    if request.method == "POST" and current_user.id != user_id:
         user_service.delete_by_id(user_id)
     return redirect("/users")
